requirements.txt
# local env madani
tensorflow==2.9.1
keras==2.9.0
Pillow==7.1.2
pandas==1.3.5
numpy==1.21.6
matplotlib==3.2.2
face_recognition==1.2.3
scikit-learn==1.0.2
tqdm==4.64.1
livelossplot==0.5.5
streamlit==1.8.1